scanner/requirements.txt
functions-framework==3.*
google-api-python-client>=2.137.0
google-auth>=2.34.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-cloud-pubsub>=2.21.0
google-cloud-secret-manager>=2.20.0
